Do not parse full config for commands that don't need it
Motivation

The [cleanup job for the Feature
Hub](https://github.com/sinnerschrader/feature-hub/actions/workflows/cleanup-stacks.yml)
(using the `purge` command) currently fails with the following Zod
validation error:

```
[
  {
    "code": "too_small",
    "minimum": 1,
    "type": "array",
    "inclusive": true,
    "exact": false,
    "message": "Array must contain at least 1 element(s)",
    "path": [
      "routes"
    ]
  }
]
```

This is because the routes are inferred from the built documentation.
But when running the `purge` command, we don't actually need the routes,
and we also don't want to build the docs just to avoid the config
validation error.

Proposed Solution

For a couple of commands (including `purge`) we only need the domain
name parts (`hostedZoneName` and sometimes `aliasRecordName`), so we can
limit the parsing/validation to those fields, and thus implicitly allow
an empty array of routes in this context.
