+++ src/cdk/create-stack.ts
@@ -1,4 +1,4 @@
-import type {StackConfig} from '../read-stack-config.js';
+import type {StackConfig} from '../parse-stack-config.js';
 
 import {getDomainName} from '../utils/get-domain-name.js';
 import {getStackName} from '../utils/get-stack-name.js';
