+++ src/dev/sort-routes.ts
@@ -1,4 +1,4 @@
-import type {Route} from '../read-stack-config.js';
+import type {Route} from '../parse-stack-config.js';
 
 export function sortRoutes<TRoute extends Pick<Route, 'publicPath'>>(
   routes: readonly TRoute[],
