+++ src/redeploy-command.ts
@@ -1,5 +1,6 @@
 import type {CommandModule} from 'yargs';
 
+import {parseDomainNameParts} from './parse-domain-name-parts.js';
 import {readStackConfig} from './read-stack-config.js';
 import {findStack} from './sdk/find-stack.js';
 import {getOutputValue} from './sdk/get-output-value.js';
@@ -36,7 +37,10 @@ export const redeployCommand: CommandModule<
 
   handler: async (args): Promise<void> => {
     const stackName =
-      args.stackName || getStackName(getDomainName(await readStackConfig()));
+      args.stackName ||
+      getStackName(
+        getDomainName(parseDomainNameParts(await readStackConfig())),
+      );
 
     print.warning(`Stack: ${stackName}`);
